Beginning of extracting household core
