+++ game/economy/households/grocery_shop/mod.rs
@@ -7,16 +7,16 @@ use economy::buildings::BuildingID;
 use economy::buildings::rendering::BuildingInspectorID;
 use transport::pathfinding::RoughLocationID;
 
-use super::{Household, HouseholdID, MemberIdx, MSG_Household_decay, MSG_Household_inspect,
-            MSG_Household_provide_deal, MSG_Household_receive_deal, MSG_Household_task_succeeded,
-            MSG_Household_task_failed, MSG_Household_destroy, MSG_Household_stop_using,
-            MSG_Household_reset_member_task};
+use super::{Household, HouseholdID, HouseholdCore, MemberIdx, MSG_Household_decay,
+            MSG_Household_inspect, MSG_Household_provide_deal, MSG_Household_receive_deal,
+            MSG_Household_task_succeeded, MSG_Household_task_failed, MSG_Household_destroy,
+            MSG_Household_stop_using, MSG_Household_reset_member_task};
 
 #[derive(Compact, Clone)]
 pub struct GroceryShop {
     id: GroceryShopID,
     site: BuildingID,
-    resources: Inventory,
+    core: HouseholdCore,
     grocery_offer: OfferID,
     job_offer: OfferID,
 }
@@ -26,7 +26,7 @@ impl GroceryShop {
         GroceryShop {
             id,
             site,
-            resources: Inventory::new(),
+            core: HouseholdCore::new(0),
             grocery_offer: OfferID::register(
                 id.into(),
                 MemberIdx(0),
@@ -51,6 +51,14 @@ impl GroceryShop {
 }
 
 impl Household for GroceryShop {
+    fn core(&self) -> &HouseholdCore {
+        &self.core
+    }
+
+    fn core_mut(&mut self) -> &mut HouseholdCore {
+        &mut self.core
+    }
+
     fn is_shared(_: Resource) -> bool {
         true
     }
